requirements.txt
opencv_python_headless>=4.2.0.32
scipy>=1.4.1
torch>=1.4.0
h5py>=2.10.0
numpy>=1.18.1
torchvision>=0.5.0
tqdm>=4.43.0
Pillow>=7.0.0
scikit_learn>=0.22.2.post1